+++ bfv/bfv_benchmark_test.go
@@ -98,19 +98,19 @@ func benchEncrypt(testctx *testContext, b *testing.B) {
 	plaintext := NewPlaintext(testctx.params)
 	ciphertext := NewCiphertextRandom(testctx.prng, testctx.params, 1)
 
-	b.Run(testString("Encrypt/key=Pk", testctx.params), func(b *testing.B) {
+	b.Run(testString("Encrypt/key=Pk/", testctx.params), func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			encryptorPk.Encrypt(plaintext, ciphertext)
 		}
 	})
 
-	b.Run(testString("Encrypt/key=Pk/Fast", testctx.params), func(b *testing.B) {
+	b.Run(testString("EncryptFast/key=Pk/", testctx.params), func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			encryptorPk.EncryptFast(plaintext, ciphertext)
 		}
 	})
 
-	b.Run(testString("Encrypt/key=Sk", testctx.params), func(b *testing.B) {
+	b.Run(testString("Encrypt/key=Sk/", testctx.params), func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			encryptorSk.Encrypt(plaintext, ciphertext)
 		}